Basic router set up and navbar moved to separate component
