+++ debug.log
@@ -1,0 +1,1 @@
+[1127/053907.921:ERROR:directory_reader_win.cc(43)] FindFirstFile: The system cannot find the path specified. (0x3)
