+++ src/pages/Bus.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+function Bus(){
+    return(
+        <h1>
+            Bus Page
+        </h1>
+    );
+}
+export default Bus;
