+++ src/pages/Cab.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+function Cab(){
+    return(
+        <h1>
+            Cab Page
+        </h1>
+    );
+}
+export default Cab;
