+++ src/pages/TimeTable.js
@@ -1,0 +1,10 @@
+import React from 'react';
+
+function TimeTable(){
+    return(
+        <h1>
+            TimeTable Page
+        </h1>
+    );
+}
+export default TimeTable;
